+++ lib/src/theme/themes/default_theme_no_secondary_border_variant.dart
@@ -9,6 +9,7 @@ import 'package:shadcn_ui/src/theme/components/accordion.dart';
 import 'package:shadcn_ui/src/theme/components/alert.dart';
 import 'package:shadcn_ui/src/theme/components/avatar.dart';
 import 'package:shadcn_ui/src/theme/components/badge.dart';
+import 'package:shadcn_ui/src/theme/components/breadcrumb.dart';
 import 'package:shadcn_ui/src/theme/components/button.dart';
 import 'package:shadcn_ui/src/theme/components/button_sizes.dart';
 import 'package:shadcn_ui/src/theme/components/calendar.dart';
@@ -234,6 +235,42 @@ class ShadDefaultThemeNoSecondaryBorderVariant extends ShadThemeVariant {
       size: const Size.square(40),
       shape: const CircleBorder(),
       backgroundColor: colorScheme.muted,
+    );
+  }
+
+  @override
+  ShadBreadcrumbTheme breadcrumbTheme() {
+    return ShadBreadcrumbTheme(
+      ellipsis: Icon(
+        LucideIcons.ellipsis,
+        size: 14,
+        color: colorScheme.mutedForeground,
+      ),
+      spacing: 10,
+      ellipsisSize: 14,
+      separatorSize: 14,
+      itemTextStyle: effectiveTextTheme.small.fallback(
+        color: colorScheme.mutedForeground,
+      ),
+      lastItemTextColor: colorScheme.foreground,
+      linkTextStyle: effectiveTextTheme.small.fallback(
+        color: colorScheme.mutedForeground,
+      ),
+      linkNormalTextColor: colorScheme.mutedForeground,
+      linkHoverTextColor: colorScheme.foreground,
+      mainAxisAlignment: WrapAlignment.start,
+      crossAxisAlignment: WrapCrossAlignment.center,
+      dropdownMenuBackgroundColor: colorScheme.popover,
+      dropdownMenuPadding: const EdgeInsets.all(4),
+      dropdownTextStyle: effectiveTextTheme.small.fallback(
+        color: colorScheme.foreground,
+      ),
+      dropdownItemPadding: const EdgeInsets.symmetric(
+        horizontal: 8,
+        vertical: 6,
+      ),
+      dropdownMenuAnchor: const ShadAnchorAuto(offset: Offset(20, 4)),
+      dropdownArrowGap: 4,
     );
   }