+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.40.0
+
+- **FEAT**: Add new component `ShadBreadcrumb` and all of its related components (thanks to @MoazSalem).
+- **FIX**: Update `ShadButton` to allow for more flexibility with height and width properties (thanks to @MoazSalem).
+
 ## 0.39.8
 
 - **FIX**: Autofocus search input in select dropdown (thanks @Isakdl).